+++ src/cli/mod.rs
@@ -1,4 +1,5 @@
 pub mod connect;
 pub mod set_key;
+pub mod stats;
 pub mod stop;
 pub mod update;